+++ src/hin/master.h
@@ -27,6 +27,7 @@ typedef struct {
   int num_active;
   uint32_t debug;
   int wait_restart;
+  int child_nr;
   //int access_fd;
   //int error_fd;
   void * servers;