+++ src/http/httpd_read.c
@@ -3,10 +3,13 @@
 #include <string.h>
 #include <stdlib.h>
 
+#include <unistd.h>
+
 #include "hin.h"
 #include "lua.h"
+#include "conf.h"
 
-int httpd_client_read_callback (hin_client_t * client, hin_buffer_t * source);
+int httpd_client_read_callback (hin_buffer_t * buffer);
 
 int httpd_client_reread (hin_client_t * client) {
   hin_buffer_t * buffer = client->read_buffer;
@@ -16,7 +19,7 @@ int httpd_client_reread (hin_client_t * client) {
   source.len = buffer->ptr - buffer->data;
   int num = 0;
   if (source.len > 0)
-    num = httpd_client_read_callback (client, buffer);
+    num = httpd_client_read_callback (buffer);
   if (num < 0) {
     printf ("client error\n");
     httpd_client_shutdown (client);
@@ -41,14 +44,15 @@ int post_done (hin_pipe_t * pipe) {
   httpd_client_finish (client);
 }
 
-int httpd_client_read_callback (hin_client_t * client, hin_buffer_t * buffer) {
+int httpd_client_read_callback (hin_buffer_t * buffer) {
   string_t source1, * source = &source1;
   source->ptr = buffer->data;
   source->len = buffer->ptr - buffer->data;
 
+  hin_client_t * client = (hin_client_t*)buffer->parent;
   httpd_client_t * http = (httpd_client_t*)&client->extra;
 
-  if (source->len > 65000) {
+  if (source->len >= HIN_HTTPD_MAX_HEADER_SIZE) {
     httpd_respond_error (client, 413, NULL);
     httpd_client_shutdown (client);
     return -1;
@@ -61,10 +65,11 @@ int httpd_client_read_callback (hin_client_t * client, hin_buffer_t * buffer) {
     off_t left = http->post_sz - consume;
     used += consume;
     // send post data in buffer to post handler
-    http->post_fd =  openat (AT_FDCWD, "/tmp", O_RDWR | O_TMPFILE, 0600);
+    http->post_fd =  openat (AT_FDCWD, HIN_HTTPD_POST_DIRECTORY, O_RDWR | O_TMPFILE, 0600);
     if (http->post_fd < 0) { printf ("openat tmpfile failed %s\n", strerror (errno)); }
     printf ("post initial %ld %ld fd %d '%.*s'\n", http->post_sz, source->len, http->post_fd, consume, source->ptr);
-    write (http->post_fd, source->ptr, consume);
+    if (write (http->post_fd, source->ptr, consume) < 0)
+      perror ("write");
     if (left > 0) {
       http->state |= HIN_REQ_POST;
       // request more post
@@ -95,6 +100,36 @@ int httpd_client_read_callback (hin_client_t * client, hin_buffer_t * buffer) {
   return used;
 }
 
+static int hin_lines_close_callback (hin_buffer_t * buffer, int ret) {
+  if (ret < 0) {
+    printf ("close fd %d had error %s\n", buffer->fd, strerror (-ret));
+    return 1;
+  }
+  hin_client_t * client = (hin_client_t*)buffer->parent;
+  hin_client_close (client);
+  return 1;
+}
+
+static int hin_lines_httpd_close_callback (hin_buffer_t * buffer) {
+  hin_client_t * client = (hin_client_t*)buffer->parent;
+  hin_lines_t * lines = (hin_lines_t*)&buffer->buffer;
+  buffer->callback = hin_lines_close_callback;
+  hin_request_close (buffer);
+  return 0;
+}
+
+hin_buffer_t * hin_lines_create (hin_client_t * client) {
+  hin_buffer_t * buf = hin_lines_create_raw ();
+  buf->fd = client->sockfd;
+  buf->parent = client;
+  buf->flags = HIN_SOCKET | (client->flags & HIN_SSL);
+  buf->ssl = &client->ssl;
+  hin_lines_t * lines = (hin_lines_t*)&buf->buffer;
+  lines->read_callback = httpd_client_read_callback;
+  lines->close_callback = hin_lines_httpd_close_callback;
+  hin_request_read (buf);
+  return buf;
+}