+++ src/netcode/buffer.c
@@ -5,11 +5,6 @@
 
 #include <hin.h>
 
-typedef struct {
-  int (*callback) (hin_client_t * client, hin_buffer_t * source);
-  int (*close_callback) (hin_buffer_t * buffer);
-} hin_lines_t;
-
 void hin_buffer_clean (hin_buffer_t * buffer) {
   if (buffer->type == HIN_DYN_BUFFER) {
     if (buffer->data) free (buffer->data);
@@ -24,6 +19,7 @@ void hin_buffer_list_remove (hin_buffer_t ** list, hin_buffer_t * new) {
     printf ("fatal error\n");
     exit (-1);
   }
+  new->next = new->prev = NULL;
 }
 
 void hin_buffer_list_append (hin_buffer_t ** list, hin_buffer_t * new) {
@@ -72,16 +68,6 @@ int hin_lines_request (hin_buffer_t * buffer) {
   hin_request_read (buffer);
 }
 
-static int hin_lines_close_callback (hin_buffer_t * buffer, int ret) {
-  if (ret < 0) {
-    printf ("close fd %d had error %s\n", buffer->fd, strerror (-ret));
-    return 1;
-  }
-  hin_client_t * client = (hin_client_t*)buffer->parent;
-  hin_client_close (client);
-  return 1;
-}
-
 static int hin_lines_read_callback (hin_buffer_t * buffer, int ret) {
   hin_lines_t * lines = (hin_lines_t*)&buffer->buffer;
   int err = 0;
@@ -95,7 +81,7 @@ static int hin_lines_read_callback (hin_buffer_t * buffer, int ret) {
   buffer->ptr += ret;
   buffer->count -= ret;
 
-  int num = lines->callback (buffer->parent, buffer);
+  int num = lines->read_callback (buffer);
   if (num < 0) {
     if (lines->close_callback) {
       err = lines->close_callback (buffer);
@@ -113,57 +99,15 @@ static int hin_lines_read_callback (hin_buffer_t * buffer, int ret) {
   return 0;
 }
 
-static int hin_lines_httpd_close_callback (hin_buffer_t * buffer) {
-  hin_client_t * client = (hin_client_t*)buffer->parent;
-  hin_lines_t * lines = (hin_lines_t*)&buffer->buffer;
-  buffer->callback = hin_lines_close_callback;
-  hin_request_close (buffer);
-  return 0;
-}
-
-hin_buffer_t * hin_lines_create (hin_client_t * client, int sockfd, int (*callback) (hin_client_t * client, hin_buffer_t * source)) {
+hin_buffer_t * hin_lines_create_raw () {
   hin_buffer_t * buf = calloc (1, sizeof *buf + sizeof (hin_lines_t));
   buf->type = HIN_DYN_BUFFER;
   buf->flags = 0;
-  buf->fd = sockfd;
-  buf->callback = hin_lines_read_callback;
   buf->count = buf->sz = READ_SZ;
   buf->pos = 0;
   buf->data = malloc (buf->sz);
   buf->ptr = buf->data;
-  buf->parent = client;
-  buf->flags = HIN_SOCKET | (client->flags & HIN_SSL);
-  buf->ssl = &client->ssl;
-  hin_lines_t * lines = (hin_lines_t*)&buf->buffer;
-  lines->callback = callback;
-  lines->close_callback = hin_lines_httpd_close_callback;
-  hin_request_read (buf);
-  return buf;
-}
-
-static int hin_lines_cgi_close_callback (hin_buffer_t * buffer) {
-  hin_client_t * client = (hin_client_t*)buffer->parent;
-  hin_lines_t * lines = (hin_lines_t*)&buffer->buffer;
-  //buffer->callback = hin_lines_close_callback;
-  //hin_request_close (buffer);
-  //printf ("cgi close headers\n");
-  return 1;
-}
-
-hin_buffer_t * hin_lines_create_cgi (hin_client_t * client, int sockfd, int (*callback) (hin_client_t * client, hin_buffer_t * source)) {
-  hin_buffer_t * buf = calloc (1, sizeof *buf + sizeof (hin_lines_t));
-  buf->type = HIN_DYN_BUFFER;
-  buf->fd = sockfd;
   buf->callback = hin_lines_read_callback;
-  buf->count = buf->sz = READ_SZ;
-  buf->pos = 0;
-  buf->data = malloc (buf->sz);
-  buf->ptr = buf->data;
-  buf->parent = client;
-  hin_lines_t * lines = (hin_lines_t*)&buf->buffer;
-  lines->callback = callback;
-  lines->close_callback = hin_lines_cgi_close_callback;
-  hin_request_read (buf);
   return buf;
 }